+++ source/robot_model/test/fixtures/generateRobotModelKinematicsTestConfigurations.m
@@ -28,6 +28,7 @@ end
 
 % initialise the random number generator
 rng(seed)
+disp(rng)
 
 % load the robot model
 robot = importrobot(urdf);
@@ -43,13 +44,20 @@ fprintf('std::vector<Eigen::MatrixXd> test_jacobian_ee_expects;\n');
 fprintf('std::vector<state_representation::CartesianPose> test_fk_ee_expects;\n');
 fprintf('std::vector<state_representation::CartesianPose> test_fk_link4_expects;\n');
 fprintf('std::vector<state_representation::CartesianTwist> test_velocity_fk_expects;\n');
+fprintf('std::vector<Eigen::MatrixXd> test_damped_jacobian_ee_expects;\n');
 fprintf('Eigen::Matrix<double, 6, 1> twist;\n');
+fprintf('Eigen::Matrix<double, 7, 1> joint_vel_damped;\n');
+fprintf('std::vector<double> test_dls_lambdas;\n');
+fprintf('std::vector<state_representation::CartesianTwist> test_ee_velocities;\n');
+fprintf('std::vector<state_representation::JointVelocities> test_velocity_damped_ik_expects;\n');
 
 % generate the configurations
 for conf = 1:nConfigurations
     % configurations
     q = robot.randomConfiguration;
     v = rand(size(q))*2 - 1;
+    v_ee = rand(6,1)*2 - 1;
+    dls_lambda = 10^(-rand());
     
     % transforms
     ee_T = robot.getTransform(q, 'panda_link8');
@@ -57,19 +65,27 @@ for conf = 1:nConfigurations
     
     % jacobian
     ee_jac = robot.geometricJacobian(q, 'panda_link8');
+    ee_jac_r(1:3,:) = ee_jac(4:6,:);
+    ee_jac_r(4:6,:) = ee_jac(1:3,:);
+    ee_jac_damped = pinv(pseudoInverseMat(ee_jac_r, dls_lambda));
     
     % expected results
     ee_pose = [tform2trvec(ee_T), tform2quat(ee_T)]; % (1x7)
     link4_pose = [tform2trvec(link4_T), tform2quat(link4_T)]; % (1x7)
     ee_twist = ee_jac * v'; % (1x6)
+    joint_vel_damped = ee_jac_damped\v_ee; % (1x7)
     
     % code generation (definitions)
     fprintf('\n// Random test configuration %i:\n', conf);
     fprintf(['state_representation::JointState config%i', ...
         '(franka->get_robot_name(), franka->get_joint_frames());\n'], ...
         conf);
-    fprintf('config%i.set_positions(%s);\n', conf, num2stdvec(q));
-    fprintf('config%i.set_velocities(%s);\n', conf, num2stdvec(v));
+    fprintf('config%i.set_positions(std::vector<double>%s);\n', conf, num2stdvec(q));
+    fprintf('config%i.set_velocities(std::vector<double>%s);\n', conf, num2stdvec(v));
+    fprintf('state_representation::CartesianTwist test_ee_velocity%i("franka");\n', conf);
+    fprintf('test_ee_velocity%i.set_data(std::vector<double>%s);\n', conf, num2stdvec(v_ee));
+    fprintf('test_ee_velocities.push_back(test_ee_velocity%i);\n', conf);
+    fprintf('test_dls_lambdas.push_back(%f);\n', dls_lambda);
     fprintf('test_configs.push_back(config%i);\n', conf);
     
     fprintf('\n// Expected results for configuration %i:\n', conf);
@@ -85,6 +101,16 @@ for conf = 1:nConfigurations
     fprintf('\t%f, %f, %f, %f, %f, %f, %f; \n', ee_jac(3, :));
     fprintf('test_jacobian_ee_expects.emplace_back(jac%i);\n', conf);
     
+    % populate damped Jacobian matrix
+    fprintf('Eigen::MatrixXd damped_jac%i(6, 7);\n', conf);
+    fprintf('damped_jac%i << %f, %f, %f, %f, %f, %f, %f, \n', conf, ee_jac_damped(4, :));
+    fprintf('\t%f, %f, %f, %f, %f, %f, %f, \n', ee_jac_damped(5, :));
+    fprintf('\t%f, %f, %f, %f, %f, %f, %f, \n', ee_jac_damped(6, :));
+    fprintf('\t%f, %f, %f, %f, %f, %f, %f, \n', ee_jac_damped(1, :));
+    fprintf('\t%f, %f, %f, %f, %f, %f, %f, \n', ee_jac_damped(2, :));
+    fprintf('\t%f, %f, %f, %f, %f, %f, %f; \n', ee_jac_damped(3, :));
+    fprintf('test_damped_jacobian_ee_expects.emplace_back(damped_jac%i);\n', conf);
+    
     % populate forward kinematics results
     fprintf(['test_fk_ee_expects.emplace_back(state_representation::CartesianPose' ...
         '("ee", Eigen::Vector3d(%f, %f, %f), Eigen::Quaterniond(%f, %f, %f, %f), ', ...
@@ -98,4 +124,9 @@ for conf = 1:nConfigurations
     fprintf('twist << %f, %f, %f, %f, %f, %f;\n', ee_twist([4:6, 1:3]));
     fprintf(['test_velocity_fk_expects.emplace_back', ...
         '(state_representation::CartesianTwist("ee", twist, franka->get_base_frame()));\n']);
+
+    % populate damped inverse velocity results
+    fprintf('joint_vel_damped << %f, %f, %f, %f, %f, %f, %f;\n', joint_vel_damped);
+    fprintf(['test_velocity_damped_ik_expects.emplace_back', ...
+        '(state_representation::JointVelocities("franka", joint_vel_damped));\n']);
 end